+++ src/routes3.cpp
@@ -11,7 +11,7 @@ void registerCleaningRoutes(crow::SimpleApp& app){
     auto normalized=trimWhitespace(parsed);
     crow::json::wvalue result;
     result["message"]="Whitespace normalized";
-    result["rows"]=normalized.size();
+    result["rows"]=(int)normalized.size();
     return crow::response(result);
   });
   CROW_ROUTE(app,"/api/standardize-case").methods("POST"_method)
@@ -35,9 +35,9 @@ void registerCleaningRoutes(crow::SimpleApp& app){
     auto parsed=parseCSV(req.body);
     auto deduped=fuzzyDeduplicateRows(parsed,threshold);
     crow::json::wvalue result;
-    result["originalRows"]=parsed.size();
-    result["deduplicatedRows"]=deduped.size();
-    result["merged"]=parsed.size()-deduped.size();
+    result["originalRows"]=(int)parsed.size();
+    result["deduplicatedRows"]=(int)deduped.size();
+    result["merged"]=(int)(parsed.size()-deduped.size());
     return crow::response(result);
   });
 }
